refactor: use celery-redbeat instead of file-based celery beat scheduler
