+++ pyproject.toml
@@ -28,6 +28,7 @@ alembic = "^1.13.1"
 pandas = "^2.2.1"
 paramiko = "^3.4.0"
 scp = "^0.14.5"
+celery-redbeat = "^2.2.0"
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^3.6.0"
